feat:  move helpers from handlers file to correct places
