+++ src/shared/lib/api/index.ts
@@ -1,9 +1,5 @@
-export * from './handlers';
 export * from './openai';
 export * from './services';
 export * from './errors';
 export * from './middleware';
-export * from './utils/summary-formatter';
-export * from './utils/ai-summary-constants';
-export * from './utils/validation-helpers';
-export * from './utils/service-factory';
+export * from './utils';
