+++ src/shared/lib/api/middleware/index.ts
@@ -1,1 +1,3 @@
 export * from './validation';
+export * from './auth';
+export * from './api-handler';
