+++ src/shared/lib/api/utils/validation-helpers.ts
@@ -8,3 +8,14 @@ export function validateNumericId(
   }
   return numericId;
 }
+
+export function validateRequiredFields(
+  body: Record<string, unknown>,
+  requiredFields: string[]
+): { isValid: boolean; missingFields: string[] } {
+  const missingFields = requiredFields.filter(field => !body[field]);
+  return {
+    isValid: missingFields.length === 0,
+    missingFields,
+  };
+}
